Extract sample data loaders for dependency reasons
